+++ handler_context_test.go
@@ -1,9 +1,10 @@
 package emperror
 
 import (
-	"errors"
 	"reflect"
 	"testing"
+
+	"emperror.dev/errors"
 )
 
 func TestHandlerContext(t *testing.T) {
@@ -12,9 +13,9 @@ func TestHandlerContext(t *testing.T) {
 	kvs := []interface{}{"a", 123}
 	handler := HandlerWith(testHandler, kvs...)
 
-	handler.Handle(errors.New("error"))
+	handler.Handle(errors.NewPlain("error"))
 
-	cerr := With(errors.New("error"), "a", 123)
+	cerr := errors.WithDetails(errors.NewPlain("error"), "a", 123)
 
 	if got, want := testHandler.LastError(), cerr; !reflect.DeepEqual(got, want) {
 		t.Errorf("error does not match the expected one\nactual:   %s\nexpected: %s", got, want)
@@ -26,9 +27,9 @@ func TestHandlerContext_Multi(t *testing.T) {
 
 	handler := HandlerWith(HandlerWith(testHandler, "a", 123), "b", 321)
 
-	handler.Handle(errors.New("error"))
+	handler.Handle(errors.NewPlain("error"))
 
-	cerr := With(errors.New("error"), "a", 123, "b", 321)
+	cerr := errors.WithDetails(errors.NewPlain("error"), "a", 123, "b", 321)
 
 	if got, want := testHandler.LastError(), cerr; !reflect.DeepEqual(got, want) {
 		t.Errorf("error does not match the expected one\nactual:   %s\nexpected: %s", got, want)
@@ -40,9 +41,9 @@ func TestHandlerContext_MultiPrefix(t *testing.T) {
 
 	handler := HandlerWithPrefix(HandlerWith(testHandler, "a", 123), "b", 321)
 
-	handler.Handle(errors.New("error"))
+	handler.Handle(errors.NewPlain("error"))
 
-	cerr := With(errors.New("error"), "b", 321, "a", 123)
+	cerr := errors.WithDetails(errors.NewPlain("error"), "b", 321, "a", 123)
 
 	if got, want := testHandler.LastError(), cerr; !reflect.DeepEqual(got, want) {
 		t.Errorf("error does not match the expected one\nactual:   %s\nexpected: %s", got, want)
@@ -54,9 +55,9 @@ func TestHandlerContext_MissingValue(t *testing.T) {
 
 	handler := HandlerWithPrefix(HandlerWith(testHandler, "k0"), "k1")
 
-	handler.Handle(errors.New("error"))
+	handler.Handle(errors.NewPlain("error"))
 
-	cerr := With(errors.New("error"), "k1", nil, "k0", nil)
+	cerr := errors.WithDetails(errors.NewPlain("error"), "k1", nil, "k0", nil)
 
 	if got, want := testHandler.LastError(), cerr; !reflect.DeepEqual(got, want) {
 		t.Errorf("error does not match the expected one\nactual:   %s\nexpected: %s", got, want)